refactor(spectral tools): SpectrumCache.cpp moved from au3 to au4
