+++ src/projectscene/view/tracksitemsview/au3/SpectrumCache.cpp
@@ -1,22 +1,16 @@
-/**********************************************************************
-
-  Audacity: A Digital Audio Editor
-
-  @file SpectrumCache.cpp
-
-  Paul Licameli split from WaveClip.cpp
-
-**********************************************************************/
+/*
+ * Audacity: A Digital Audio Editor
+ */
 
 #include "SpectrumCache.h"
-
-#include "SpectrogramSettings.h"
-#include "RealFFTf.h"
-#include "Sequence.h"
-#include "Spectrum.h"
 #include "WaveClipUIUtilities.h"
-#include "WaveTrack.h"
-#include "WideSampleSequence.h"
+
+#include "libraries/lib-fft/RealFFTf.h"
+#include "libraries/lib-fft/Spectrum.h"
+#include "libraries/lib-mixer/WideSampleSequence.h"
+#include "libraries/lib-wave-track-settings/SpectrogramSettings.h"
+#include "libraries/lib-wave-track/Sequence.h"
+#include "libraries/lib-wave-track/WaveTrack.h"
 #include <cmath>
 
 namespace {
